finish TermsAndProccess component ; new Submit component init ; more updates
