+++ src/components/Vulnerability/Submit/Submit.tsx
@@ -1,0 +1,44 @@
+import React from "react";
+import { useSelector } from "react-redux";
+import { ICardData } from "../VulnerabilityAccordion";
+import { isProviderAndNetwork } from "../../../utils";
+import { RootState } from "../../../reducers";
+import SubmitSuccess from "./components/SubmitSuccess/SubmitSuccess";
+import SubmitReview from "./components/SubmitReview/SubmitReview";
+import "./index.scss";
+
+interface IProps {
+  cards: { [id: number]: ICardData }
+  setCards: Function
+}
+
+export default function Submit({ cards, setCards }: IProps) {
+
+
+  return (
+    <div className="submit-wrapper card-content">
+      {cards[5].verified ? <SubmitSuccess />
+
+
+
+        // (
+        //   <>
+        //     <div>Hey new Hatter, thank you for your submission!</div>
+        //     <div>Become an active member of our community and let us know what you think.</div>
+        //     <div className="submit-complete-actions-wrapper">
+        //       <button onClick={() => { props.setCards(VULNERABILITY_INIT_DATA); toggleCard(1) }}>SUBMIT NEW VULNERABILITY</button>
+        //       <button className="discord-btn" onClick={() => window.open(SocialLinks.Discord)}>
+        //         <span>JOIN US ON DISCORD</span>
+        //         <DiscordIcon fill={Colors.darkBlue} />
+        //       </button>
+        //     </div>
+        //   </>) 
+
+        :
+
+        <SubmitReview cards={cards} />
+
+      }
+    </div>
+  )
+}
