+++ src/components/Vulnerability/Submit/components/SubmitReview/index.scss
@@ -1,0 +1,13 @@
+@import "../../../../../styles/constants.scss";
+
+.submit-review-wrapper {
+  display: flex;
+  flex-direction: column;
+  button {
+    min-width: 150px;
+  }
+  .error-label {
+    color: $red;
+    margin: 10px 0px;
+  }
+}
