+++ src/components/Vulnerability/Submit/components/SubmitSuccess/SubmitSuccess.tsx
@@ -1,0 +1,10 @@
+
+import "./index.scss";
+
+export default function SubmitSuccess() {
+  return (
+    <div className="submit-success-wrapper">
+      SubmitSuccess
+    </div>
+  )
+}
