+++ src/components/Vulnerability/Submit/index.scss
@@ -1,26 +1,8 @@
-@import "../constants.scss";
+@import "../../../styles/constants.scss";
 
 .submit-wrapper {
   position: relative;
 
-  .submit-action-wrapper {
-    display: flex;
-    align-items: center;
-    button {
-      min-width: 150px;
-    }
-    input {
-      width: 20px;
-      height: 20px;
-    }
-    label {
-      margin-left: 10px;
-    }
-  }
-  .error-label {
-    color: $red;
-    margin: 10px 0px;
-  }
 
   &.loading {
     opacity: 0.5;
@@ -53,12 +35,7 @@
     .total-vault-wrapper {
       font-size: $xxsmall;
     }
-    .submit-action-wrapper {
-      flex-direction: column;
-      button {
-        margin-top: 30px;
-      }
-    }
+
     .submit-complete-actions-wrapper {
       flex-direction: column;
       > button {
