+++ src/components/Vulnerability/TermsAndProcess/TermsAndProcess.tsx
@@ -12,6 +12,11 @@ import { useSelector } from "react-redux";
 import { RootState } from "../../../reducers";
 import Modal from "../../Shared/Modal";
 import NFTPrize from "../../NFTPrize";
+import QuestionIcon from "../../../assets/icons/question-icon.svg";
+import ErrosIcon from "../../../assets/icons/error-icon.svg";
+import TimelineIcon from "../../../assets/icons/timeline-icon.svg";
+import BugIcon from "../../../assets/icons/bug-icon.svg";
+import RewardIcon from "../../../assets/icons/reward-icon.svg";
 import "./index.scss";
 
 interface IProps {
@@ -61,7 +66,35 @@ export default function TermsAndProcess({ cards }: IProps) {
       <div className="section-content submission">{t("SubmitVulnerability.TermsAndProcess.submission")}</div>
 
       <div className="section-title fix">2 {t("SubmitVulnerability.TermsAndProcess.fix-sub-title")}</div>
-      <div className="section-content fix">{t("SubmitVulnerability.TermsAndProcess.fix")}</div>
+      <div className="section-content fix">
+        {t("SubmitVulnerability.TermsAndProcess.fix")}
+        <div className="icon-text-wrapper">
+          <img src={QuestionIcon} alt="question icon" />
+          {t("SubmitVulnerability.TermsAndProcess.fix-question")}
+        </div>
+
+        <div className="icon-text-wrapper">
+          <img src={ErrosIcon} alt="error icon" />
+          {t("SubmitVulnerability.TermsAndProcess.fix-error")}
+        </div>
+
+        {t("SubmitVulnerability.TermsAndProcess.fix-text-1")}
+
+        <div className="icon-text-wrapper">
+          <img src={TimelineIcon} alt="timeline icon" />
+          {t("SubmitVulnerability.TermsAndProcess.fix-timeline")}
+        </div>
+
+        <div className="icon-text-wrapper">
+          <img src={BugIcon} alt="bugs icon" />
+          {t("SubmitVulnerability.TermsAndProcess.fix-bugs")}
+        </div>
+
+        <div className="icon-text-wrapper">
+          <img src={RewardIcon} alt="rewards icon" />
+          {t("SubmitVulnerability.TermsAndProcess.fix-rewards")}
+        </div>
+      </div>
 
       <div className="section-title awards">3 {t("SubmitVulnerability.TermsAndProcess.awards-sub-title")}</div>
       <div className="section-content awards">
@@ -77,7 +110,7 @@ export default function TermsAndProcess({ cards }: IProps) {
               {prizes}
             </tbody>
           </table>) : "Please choose project to view prizes"}
-          {t("SubmitVulnerability.TermsAndProcess.awards-text-2")}
+        {t("SubmitVulnerability.TermsAndProcess.awards-text-2")}
       </div>
 
       <div className="warning-notice">{t("SubmitVulnerability.TermsAndProcess.warning-notice")}</div>
