+++ src/components/Vulnerability/TermsAndProcess/index.scss
@@ -6,7 +6,7 @@
     width: 100%;
     border-collapse: collapse;
     margin-bottom: 15px;
-    
+
     tr {
       th {
         text-align: left;
@@ -79,6 +79,15 @@
     }
     &.fix {
       border-color: $red;
+      .icon-text-wrapper {
+        display: flex;
+        align-items: center;
+        margin: 30px 0;
+        padding-left: 15px;
+        img {
+          margin-right: 10px;
+        }
+      }
     }
     &.awards {
       border-color: $dirty-turquoise;
@@ -86,7 +95,7 @@
   }
   .warning-notice {
     border: 1px solid $red;
-    padding: 10px 30px;
+    padding: 30px 80px;
     margin: 30px 0;
   }
   .accept-terms-wrapper {
@@ -96,7 +105,7 @@
       min-width: 150px;
     }
     .checkbox-container {
-      input[type='checkbox']{
+      input[type="checkbox"] {
         margin-right: 10px;
       }
     }
@@ -105,10 +114,14 @@
 
 @media only screen and (max-width: $mobile-breakpoint) {
   .terms-and-process-wrapper {
-    tr, td {
+    tr,
+    td {
       padding: 10px;
     }
     .section-content {
+      padding: 10px;
+    }
+    .warning-notice {
       padding: 10px;
     }
     .accept-terms-wrapper {
