+++ src/components/Vulnerability/VulnerabilityAccordion.tsx
@@ -5,7 +5,7 @@ import ProjectSelect from "./ProjectSelect";
 import Description from "./Description";
 import ContactInfo from "./ContactInfo";
 import TermsAndProcess from "./TermsAndProcess/TermsAndProcess";
-import Submit from "./Submit";
+import Submit from "./Submit/Submit";
 import { getAppVersion } from "../../utils";
 import { useSelector } from "react-redux";
 import { RootState } from "../../reducers";
